decouple llm config from task generator
