+++ deval/llms/openai_llm.py
@@ -1,7 +1,8 @@
 
 import time
 import bittensor as bt
-from deval.llms.base_llm import BaseLLM, LLMArgs, LLMAPIs
+from deval.llms.base_llm import BaseLLM
+from deval.llms.llm_config import LLMAPIs, LLMArgs
 from openai import OpenAI 
 import os
