+++ deval/rewards/reward.py
@@ -176,28 +176,26 @@ class BaseRewardModel(ABC):
 
 if __name__ == "__main__":
     from deval.tasks import TasksEnum
-    from deval.task_generator import create_task
+    from deval.task_generator import TaskGenerator
     from deval.agent import HumanAgent
     from deval.protocol import EvalSynapse
     from deval.dendrite import DendriteResponseEvent
     from deval.rewards.pipeline import RewardPipeline
-    from deval.llms.openai_llm import OpenAILLM
-    from deval.llms.base_llm import LLMArgs, LLMFormatType
+    from deval.llms.llm_config import LLMAPIs
     from dotenv import load_dotenv, find_dotenv
     
     task_name = TasksEnum.HALLUCINATION.value
     _ = load_dotenv(find_dotenv())
 
-    model_kwargs = LLMArgs(format = LLMFormatType.TEXT)
-    llm_pipeline = OpenAILLM(
-        model_id="gpt-4o-mini",
-        model_kwargs=model_kwargs
-    )
- 
+    task_generator = TaskGenerator()
 
-    # get task 
-    task_name = TasksEnum.COMPLETENESS.value
-    task = create_task(llm_pipeline, task_name)
+    llm_pipeline = [
+        model for model in task_generator.available_models 
+        if model.api == LLMAPIs.OPENAI 
+        and model.model_id == "gpt-4o-mini"
+    ][0]
+ 
+    task = task_generator.create_task(llm_pipeline, task_name)
     agent = HumanAgent(task=task)
 
     # prep fake response
